+++ components/collector/tests/source_collectors/sonarqube/test_commented_out_code.py
@@ -1,5 +1,7 @@
 """Unit tests for the SonarQube commented-out code collector."""
 
+from shared_data_model import DATA_MODEL
+
 from .base import SonarQubeTestCase
 
 
@@ -12,12 +14,10 @@ class SonarQubeCommentedOutCodeTest(SonarQubeTestCase):
         """Test that the number of lines with commented out code is returned."""
         json = {"total": "2"}
         response = await self.collect(get_request_json_return_value=json)
+        expected_rules = ",".join(sorted(DATA_MODEL.sources["sonarqube"].configuration["commented_out_rules"].value))
         self.assert_measurement(
             response,
             value="2",
             total="100",
-            landing_url=f"{self.issues_landing_url}&rules=abap:S125,c:CommentedCode,cpp:CommentedCode,"
-            "flex:CommentedCode,csharpsquid:S125,javascript:S125,kotlin:S125,"
-            "objc:CommentedCode,php:S125,plsql:S125,python:S125,scala:S125,"
-            "java:S125,swift:S125,typescript:S125,Web:AvoidCommentedOutCodeCheck,xml:S125",
+            landing_url=f"{self.issues_landing_url}&rules={expected_rules}",
         )