+++ components/collector/tests/source_collectors/sonarqube/test_long_units.py
@@ -1,5 +1,7 @@
 """Unit tests for the SonarQube long units collector."""
 
+from shared_data_model import DATA_MODEL
+
 from .base import SonarQubeTestCase
 
 
@@ -23,14 +25,10 @@ class SonarQubeLongUnitsTest(SonarQubeTestCase):
                 long_units_json,
             ],
         )
+        expected_rules = ",".join(sorted(DATA_MODEL.sources["sonarqube"].configuration["long_unit_rules"].value))
         self.assert_measurement(
             response,
             value="2",
             total="4",
-            landing_url=f"{self.issues_landing_url}&rules=abap:S104,c:FileLoc,cpp:FileLoc,csharpsquid:S104,"
-            "csharpsquid:S138,flex:S138,go:S104,go:S138,javascript:S104,javascript:S138,kotlin:S104,"
-            "kotlin:S138,objc:FileLoc,php:S104,php:S138,php:S2042,Pylint:R0915,python:S104,python:S138,"
-            "ruby:S104,ruby:S138,scala:S104,scala:S138,java:S104,java:S1188,java:S138,java:S2972,"
-            "swift:S104,typescript:S104,typescript:S138,vbnet:S104,vbnet:S138,Web:FileLengthCheck,"
-            "Web:LongJavaScriptCheck",
+            landing_url=f"{self.issues_landing_url}&rules={expected_rules}",
         )